FindWordsModule: Split up into two classes
This splits the module-related code from the actual algorithm.
