+++ src/module/uniol/apt/analysis/synthesize/FindWordsModule.java
@@ -1,6 +1,6 @@
 /*-
  * APT - Analysis of Petri Nets and labeled Transition systems
- * Copyright (C) 2014-2015  Uli Schlachter
+ * Copyright (C) 2014-2016  Uli Schlachter
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -21,16 +21,10 @@ package uniol.apt.analysis.synthesize;
 
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.Iterator;
 import java.util.List;
-import java.util.Map;
 import java.util.SortedSet;
 import java.util.TreeSet;
 
-import uniol.apt.adt.ts.TransitionSystem;
-import uniol.apt.analysis.exception.NonDeterministicException;
 import uniol.apt.module.AbstractModule;
 import uniol.apt.module.AptModule;
 import uniol.apt.module.Category;
@@ -40,7 +34,7 @@ import uniol.apt.module.ModuleInputSpec;
 import uniol.apt.module.ModuleOutput;
 import uniol.apt.module.ModuleOutputSpec;
 import uniol.apt.module.exception.ModuleException;
-import static uniol.apt.analysis.synthesize.SynthesizeUtils.*;
+import uniol.apt.analysis.synthesize.SynthesizeUtils;
 
 /**
  * Find all words that can be generated by a Petri net from a given class.
@@ -48,14 +42,6 @@ import static uniol.apt.analysis.synthesize.SynthesizeUtils.*;
  */
 @AptModule
 public class FindWordsModule extends AbstractModule implements Module {
-	public interface WordCallback {
-		public void call(List<String> wordAsList, String wordAsString, SynthesizePN synthesize);
-	}
-
-	public interface LengthDoneCallback {
-		public void call(int length);
-	}
-
 	static private enum Operation {
 		UNSOLVABLE(true, true, false),
 		SOLVABLE(true, false, true),
@@ -167,7 +153,7 @@ public class FindWordsModule extends AbstractModule implements Module {
 		// Indices into the above array
 		final int solvable = 0;
 		final int unsolvable = 1;
-		WordCallback wordCallback = new WordCallback() {
+		FindWords.WordCallback wordCallback = new FindWords.WordCallback() {
 			@Override
 			public void call(List<String> wordAsList, String wordAsString, SynthesizePN synthesize) {
 				if (synthesize.wasSuccessfullySeparated()) {
@@ -177,13 +163,13 @@ public class FindWordsModule extends AbstractModule implements Module {
 				} else {
 					counters[unsolvable]++;
 					if (printUnsolvable)
-						System.out.println(formatESSPFailure(wordAsList,
+						System.out.println(SynthesizeUtils.formatESSPFailure(wordAsList,
 									synthesize.getFailedEventStateSeparationProblems(),
 									true));
 				}
 			}
 		};
-		LengthDoneCallback lengthDoneCallback = new LengthDoneCallback() {
+		FindWords.LengthDoneCallback lengthDoneCallback = new FindWords.LengthDoneCallback() {
 			@Override
 			public void call(int length) {
 				System.out.println("Done with length " + length + ". There were " + counters[unsolvable]
@@ -192,87 +178,7 @@ public class FindWordsModule extends AbstractModule implements Module {
 				counters[unsolvable] = 0;
 			}
 		};
-		generateList(properties, alphabet, !printUnsolvable, wordCallback, lengthDoneCallback);
-	}
-
-	static public void generateList(PNProperties properties, SortedSet<String> alphabet, boolean quickFail, WordCallback wordCallback, LengthDoneCallback lengthDoneCallback) {
-		List<String> currentLevel = Collections.singletonList("");
-		List<String> nextLevel = new ArrayList<>();
-
-		while (!currentLevel.isEmpty()) {
-			for (String currentWord : currentLevel) {
-				for (String c : alphabet) {
-					boolean newLetter = !currentWord.contains(c);
-
-					// If "currentWord" is unsolvable, then "word" must also be unsolvable.
-					// Otherwise we get a contradiction: The net solving "word" will solve
-					// "currentWord" after firing "c" once.
-					// Put differently: By prepending letters to solvable words, we are sure to
-					// generate all solvable words.
-					String word = c + currentWord;
-					word = normalizeWord(toList(word), alphabet);
-					List<String> wordList = toList(word);
-
-					if (!properties.isKBounded()) {
-						// If we have unbounded places, then every prefix of a solvable word is
-						// also solvable: Just add a place from which every transition consumes
-						// one token. This place's initial marking limits the length of the
-						// word.
-						// For our purpose this means: If the prefix isn't solvable, then we
-						// already know that the word itself isn't solvable either.
-						// This is also important for the definition of "minimally unsolvable"
-						// (= unsolvable + all proper subwords are solvable).
-						if (Collections.binarySearch(currentLevel, normalizeWord(
-								wordList.subList(0, wordList.size() - 1), alphabet)) < 0)
-							continue;
-					} else {
-						// In a k-bounded Petri net, every suffix of a solvable word is also
-						// solvable. However, a prefix might still be unsolvable. Thus, we use a
-						// different definition of "minimally unsolvable" here (= unsolvable +
-						// all proper suffixes are solvable).
-					}
-
-					// Is "word" PN-solvable with the given properties?
-					TransitionSystem ts = makeTS(wordList);
-					SynthesizePN synthesize;
-					try {
-						synthesize = new SynthesizePN.Builder(ts)
-							.setProperties(properties)
-							// we don't need failed separation points, if we don't show them
-							.setQuickFail(quickFail)
-							.buildForLanguageEquivalence();
-					} catch (MissingLocationException e) {
-						throw new RuntimeException("Not generating locations and "
-								+ " yet they were generated wrongly?!", e);
-					} catch (NonDeterministicException e) {
-						throw new RuntimeException("Generated a deterministic TS and "
-								+ " yet it is non-deterministic?!", e);
-					}
-					wordCallback.call(wordList, word, synthesize);
-					if (synthesize.wasSuccessfullySeparated()) {
-						nextLevel.add(word);
-					}
-
-					if (newLetter)
-						// The alphabet is a sorted set. We only extend words in the order that
-						// they appear in the alphabet. So if the current letter was new, then
-						// all the following ones will be new, too.
-						// When extending "ba", "cab" is solvable if and only if "dab" is
-						// solvable. So trying other new letters won't produce really "new"
-						// words, but only words that are symmetric in the sense that they can
-						// be transformed into each other by replacing one letter with another.
-						// Avoiding these symmetries in the words we generate helps speeding up
-						// this algorithm.
-						break;
-				}
-			}
-
-			int currentLength = currentLevel.iterator().next().length() + 1;
-			lengthDoneCallback.call(currentLength);
-			currentLevel = nextLevel;
-			nextLevel = new ArrayList<>();
-			Collections.sort(currentLevel);
-		}
+		FindWords.generateList(properties, alphabet, !printUnsolvable, wordCallback, lengthDoneCallback);
 	}
 
 	// Transform a String into the list of its characters
@@ -281,26 +187,6 @@ public class FindWordsModule extends AbstractModule implements Module {
 		// some Java versions include "" as first part of the splitted string => try to remove it
 		result.remove("");
 		return result;
-	}
-
-	// Normalize a word into the form that the above loop would generate it in. This means e.g. that the word ends
-	// with the first letter of the alphabet.
-	static private String normalizeWord(List<String> word, SortedSet<String> alphabet)
-	{
-		StringBuilder result = new StringBuilder();
-		Map<String, String> morphism = new HashMap<>();
-		Iterator<String> alphabetIter = alphabet.iterator();
-
-		for (String letter : word) {
-			String replacement = morphism.get(letter);
-			if (replacement == null) {
-				assert alphabetIter.hasNext();
-				replacement = alphabetIter.next();
-				morphism.put(letter, replacement);
-			}
-			result.append(replacement);
-		}
-		return result.toString();
 	}
 
 	@Override
